support SQL WHERE semantics, finished adding tests for skipping logic
